+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.1.1
+
+- Relic Draws
+  - Add anima lens information to relic draw selections
+
 # v2.1.0
 
 - New feature: Export